+++ src/test/java/seedu/address/logic/commands/event/EditEventCommandTest.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.event;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
@@ -19,6 +19,7 @@ import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.CommandHistory;
 import seedu.address.logic.UndoRedoStack;
+import seedu.address.logic.commands.ClearCommand;
 import seedu.address.logic.commands.event.EditEventCommand;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
