+++ src/test/java/seedu/address/logic/commands/imports/ImportCommandTest.java
@@ -1,0 +1,4 @@
+package seedu.address.logic.commands.imports;
+
+public class ImportCommandTest {
+}
